+++ alacritty/Cargo.toml
@@ -10,10 +10,14 @@ edition = "2018"
 
 [dependencies]
 alacritty_terminal = { path = "../alacritty_terminal" }
+clap = "2"
 log = "0.4"
 time = "0.1.40"
 env_logger = "0.6.0"
 crossbeam-channel = "0.3.8"
+
+[build-dependencies]
+rustc_tools_util = "0.1"
 
 [target.'cfg(target_os = "macos")'.dependencies]
 dirs = "1.0.2"
